BINIUS-412: reduce the monolithic and batched verifiers through one shared driver
Two verifiers reduced the same ConstraintSystem through duplicated logic. The
batched one was the monolithic one with log_instances added to every row count,
plus one step transporting each operation's operand claims onto a shared
instance point. Every ordering invariant, skip guard and domain choice was
written twice, and a chip architecture would have made it three times.

Extract one driver parameterized by the instance count. A batch of one instance
is the monolithic reduction, so the transport step is guarded on there being
more than one instance point to unify:

    transport  <=>  log_instances > 0 and a multiplication is present

With one instance every operation's instance point is empty, so there is nothing
to transport. That guard is what keeps the monolithic transcript unchanged.

Both verifiers shrink to: receive the trace oracle, reduce, ring-switch, open.
The reduction output carries the trace opening point, so neither caller
assembles r_j || r_rho || r_y by hand. The two drivers lose 543 lines.

The batched prover's guard gains the same instance condition. Its no-transport
branch previously assumed no multiplications existed and zeroed their claims,
which a single instance can now reach; a present multiplication must still
contribute its claim, collapsed from its per-bit form.
